# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditEventKind": {
        "enum": [
          "grant_minted",
          "grant_revoked",
          "emergency_exception",
          "injection_blocked",
          "model_output_rejected",
          "write_executed",
          "write_replayed",
          "write_compensated",
          "notification_sent",
          "approval_granted",
          "circuit_opened",
          "circuit_closed",
          "rms_flushed",
          "dead_lettered"
        ],
        "title": "AuditEventKind",
        "type": "string"
      },
      "AuditEventView": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Detail",
            "type": "object"
          },
          "kind": {
            "$ref": "#/components/schemas/AuditEventKind"
          },
          "occurred_at": {
            "title": "Occurred At",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "audit_id",
          "kind",
          "occurred_at",
          "actor",
          "correlation_id",
          "detail"
        ],
        "title": "AuditEventView",
        "type": "object"
      },
...
      "DeadLetterListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "dead_letters": {
            "items": {
              "$ref": "#/components/schemas/DeadLetterView"
            },
            "title": "Dead Letters",
            "type": "array"
          }
        },
        "required": [
          "dead_letters",
          "count"
        ],
        "title": "DeadLetterListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyDecisionView": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "decided_at": {
            "title": "Decided At",
            "type": "string"
          },
          "decided_by": {
            "title": "Decided By",
            "type": "string"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "justification": {
            "title": "Justification",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "decision_id",
          "agent_id",
          "target",
          "operation",
          "classification",
          "action",
          "rule_id",
          "justification",
          "policy_version",
          "decided_at",
          "decided_by"
        ],
        "title": "PolicyDecisionView",
        "type": "object"
      },
      "PushResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "deliveries": {
            "items": {
              "$ref": "#/components/schemas/DeliveryReport"
            },
            "title": "Deliveries",
            "type": "array"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "poison_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poison Reason"
          },
          "topic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Topic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "accepted"
        ],
        "title": "PushResponse",
        "type": "object"
      },
...
      "SchedulerTickRequest": {
        "additionalProperties": false,
        "description": "What Cloud Scheduler sends on each tick.",
        "properties": {
          "district_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "reason": {
            "default": "scheduled",
            "maxLength": 60,
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "SchedulerTickRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Municipal structural intelligence as an institutional agent fleet.\n\nA decision-support prototype. It delivers information and performs clerical\nexecution. It never issues tactical recommendations, evacuation orders, crew\nassignments, or fire-behaviour predictions.",
    "title": "FIRST DUE",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/internal/audit/decisions": {
      "get": {
        "description": "Why every access was allowed, derived, withheld, staged, or denied.",
        "operationId": "list_policy_decisions_api_v1_internal_audit_decisions_get",
        "parameters": [
          {
            "in": "query",
            "name": "incident_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicyDecisionView"
                  },
                  "title": "Response List Policy Decisions Api V1 Internal Audit Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gateway policy decisions",
        "tags": [
          "internal"
        ]
      }
    },
    "/api/v1/internal/audit/events": {
      "get": {
        "description": "Every recorded decision, exception, block, write, and grant.\n\nDetail was redacted on the way in, so this endpoint cannot leak what the\nrecords said even to a caller who is entitled to read the audit log.",
        "operationId": "list_audit_events_api_v1_internal_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "incident_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditEventKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "correlation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Correlation Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventView"
                  },
                  "title": "Response List Audit Events Api V1 Internal Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Immutable audit events",
        "tags": [
          "internal"
        ]
      }
    },
    "/api/v1/internal/callbacks/write": {
      "post": {
        "description": "Accept an acknowledgement from a system we wrote to.\n\nAuthenticated by HMAC signature over the method, path, timestamp, and a hash\nof the body -- not by the caller asserting who it is. The timestamp is\ninside the signed material and carries a freshness window, because a\nsignature that never goes stale is a replay waiting to happen.\n\nA callback changes what a profile says about a filed referral, so a forged\none would be a way to write to the record from outside.",
        "operationId": "write_callback_api_v1_internal_callbacks_write_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Write Callback Api V1 Internal Callbacks Write Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, stale, or invalid signature."
          }
        },
        "summary": "Signed callback from a receiving system",
        "tags": [
          "internal"
        ]
      }
    },
    "/api/v1/internal/events/dead-letters": {
      "get": {
        "description": "Everything the fleet could not process. Surfaced, never dropped.",
        "operationId": "list_dead_letters_api_v1_internal_events_dead_letters_get",
        "parameters": [
          {
            "in": "query",
            "name": "subscriber",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subscriber"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dead-lettered envelopes",
        "tags": [
          "internal"
        ]
      }
    },
    "/api/v1/internal/events/push": {
      "post": {
        "description": "Accept one pushed envelope and route it to local subscribers.",
        "operationId": "push_event_api_v1_internal_events_push_post",
        "parameters": [
          {
            "in": "query",
            "name": "subscriber",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subscriber"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Not processed; the broker should redeliver."
          }
        },
        "summary": "Pub/Sub push delivery",
        "tags": [
          "internal"
        ]
      }
    },
...
    "/api/v1/internal/scheduler/tick": {
      "post": {
        "description": "Drive the slow loop on a schedule.\n\nPhase 3 built the loop and phase 3's own notes recorded the gap: \"`poll`\nruns when something calls it. Cloud Scheduler wiring is not written.\" This\nis what Cloud Scheduler calls, authenticated as a service the same way the\nPub/Sub push endpoint is.\n\nA referral is never auto-approved here. The pass stages it and a human\nfiles it, exactly as it does when a person triggers the poll.",
        "operationId": "scheduler_tick_api_v1_internal_scheduler_tick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulerTickRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Scheduler Tick Api V1 Internal Scheduler Tick Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cloud Scheduler tick: run one slow-loop pass",
        "tags": [
          "internal"
        ]
      }
    },
...
  }
}